+++ elastiknn-client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/ElastiknnClient.scala
@@ -8,10 +8,10 @@ import com.sksamuel.elastic4s.requests.bulk.{BulkResponse, BulkResponseItem}
 import com.sksamuel.elastic4s.requests.indexes.{CreateIndexResponse, PutMappingResponse}
 import com.sksamuel.elastic4s.requests.searches.{SearchRequest, SearchResponse}
 import org.apache.http.HttpHost
-import org.apache.http.client.config.RequestConfig
 import org.elasticsearch.client.RestClient
 
 import scala.annotation.tailrec
+import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
 import scala.language.higherKinds
 
@@ -100,14 +100,15 @@ object ElastiknnClient {
 
   final case class StrictFailureException(message: String, cause: Throwable = None.orNull) extends RuntimeException(message, cause)
 
-  def futureClient(host: String = "localhost", port: Int = 9200, strictFailure: Boolean = true, timeoutMillis: Int = 30000)(
-      implicit ec: ExecutionContext
-  ): ElastiknnFutureClient = {
-    val rc: RestClient = RestClient
-      .builder(new HttpHost(host, port))
-      .setRequestConfigCallback((requestConfigBuilder: RequestConfig.Builder) => requestConfigBuilder.setSocketTimeout(timeoutMillis))
-      .build()
-    val jc: JavaClient = new JavaClient(rc)
+  /**
+    * Build an [[ElastiknnFutureClient]] from an elasticsearch RestClient.
+    * @param restClient The Elasticsearch RestClient, configured how you like, e.g. connected to multiple nodes.
+    * @param strictFailure If true, convert non-200 responses and bulk responses containing any failure to a failed Future.
+    * @param ec ExecutionContext where requests are executed and responses processed.
+    * @return [[ElastiknnFutureClient]]
+    */
+  def futureClient(restClient: RestClient, strictFailure: Boolean)(implicit ec: ExecutionContext): ElastiknnFutureClient = {
+    val jc: JavaClient = new JavaClient(restClient)
     new ElastiknnFutureClient {
       implicit val executor: Executor[Future] = Executor.FutureExecutor(ec)
       implicit val functor: Functor[Future] = Functor.FutureFunctor(ec)
@@ -122,10 +123,29 @@ object ElastiknnClient {
         }
         else future
       }
-
-      override def toString: String = s"${ElastiknnClient.getClass.getSimpleName} connected to $host:$port"
+      override def toString: String =
+        s"${ElastiknnClient.getClass.getSimpleName} connected to ${restClient.getNodes.asScala.toList.mkString(",")}"
       override def close(): Unit = elasticClient.close()
     }
+  }
+
+  /**
+    * Build an [[ElastiknnFutureClient]] that connects to a single host
+    * @param host Elasticsearch host.
+    * @param port Elasticsearch port.
+    * @param timeoutMillis Amount of time to wait for a response.
+    * @param strictFailure If true, convert non-2xx responses and bulk responses containing any failure to a failed Future.
+    * @param ec ExecutionContext where requests are executed and responses processed.
+    * @return [[ElastiknnFutureClient]]
+    */
+  def futureClient(host: String = "localhost", port: Int = 9200, strictFailure: Boolean = true, timeoutMillis: Int = 30000)(
+      implicit ec: ExecutionContext
+  ): ElastiknnFutureClient = {
+    val rc: RestClient = RestClient
+      .builder(new HttpHost(host, port))
+      .setRequestConfigCallback(_.setSocketTimeout(timeoutMillis))
+      .build()
+    futureClient(rc, strictFailure)
   }
 
   private def checkResponse[T, U](req: T, res: Response[U]): Either[Throwable, U] = {